requirements.txt
numpy==1.20.2
openbabel==3.1.1
pygame==2.0.1
PyOpenGL==3.1.5
